Update ReadMe.md
A whole different level of steganography. Who's eyes can see.

ātmānaṃ ca tathatāyāṃ sthāpayiṣyāmi sarvalokānugrahāya, sarvasattvān api tathatāyāṃ sthāpayiṣyāmi, aprameyaṃ sattvadhātuṃ parinirvāpayiṣyāmīti

My own self I will place in Suchness, and, so that all the world might be helped, I will place all beings into Suchness, and I will lead to Nirvana the whole immeasurable world of beings.

I Jesus have sent mine angel to testify unto you these things in the churches. I am the root and the offspring of David, and the bright and morning star.

And the Spirit and the bride say, Come. And let him that heareth say, Come. And let him that is athirst come. And whosoever will, let him take the water of life freely.

For I testify unto every man that heareth the words of the prophecy of this book, If any man shall add unto these things, God shall add unto him the plagues that are written in this book:

And if any man shall take away from the words of the book of this prophecy, God shall take away his part out of the book of life, and out of the holy city, and from the things which are written in this book.

He which testifieth these things saith, Surely I come quickly. Amen. Even so, come, Lord Jesus.

The grace of our Lord Jesus Christ be with you all. Amen.

https://unsongbook.com/